requirements.txt
asgiref==3.5.2
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.5
Django==4.0.6
django-environ==0.9.0
django-import-export==2.8.0
et-xmlfile==1.1.0
filelock==3.8.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.10
Pillow==9.2.0
platformdirs==2.5.2
PyYAML==6.0
sqlparse==0.4.2
tablib==3.2.1
tzdata==2022.1
virtualenv==20.16.3
xlrd==2.0.1
xlwt==1.3.0
